tinycompress: do not loop in the compress read
Return the compress read immediately when it is
successful. client can recall for the remaining bytes.
Also, In case of capture, move the plugin state
to prepared when set params is successful.
Earlier in case of playback this plug state is moved
to prepare only when first write is successful.

Change-Id: I9a6a365d78ede8ff04e4d99bcdaca05a71831d6c
CRs-Fixed: 3424064